+++ addon-test-support/-private/steps/then.js
@@ -4,7 +4,7 @@ import { expect } from 'chai';
 import { assert }  from '@ember/debug';
 import { currentURL, pauseTest, settled } from '@ember/test-helpers';
 import { isVisible, pause } from 'ember-cli-yadda-opinionated/test-support/-private/helpers';
-import { findRadioForLabelWithText } from 'ember-cli-yadda-opinionated/test-support/-private/dom-helpers';
+import { findInputForLabelWithText } from 'ember-cli-yadda-opinionated/test-support/-private/dom-helpers';
 import { camelize }  from '@ember/string';
 import { pluralize } from 'ember-inflector';
 
@@ -126,14 +126,36 @@ const steps = {
     expect(record[key]).deep.equal(value);
   },
 
-  "Then radio button \"(.+?)\" should be selected in $opinionatedElement"(text, [collection/* , label, selector */]) {
+  "Then (?:the )?(?:radio button|checkbox) $opinionatedElement should (NOT |not )?be selected"([collection/* , label, selector */], not) {
     assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
     const [element] = collection;
-    const radioButton = findRadioForLabelWithText(element, text);
+    let input;
 
-    return radioButton.checked;
-  }
+    if (element.type === 'checkbox' || element.type === 'radio') {
+      input = element;
+    } else {
+      const inputs = element.querySelectorAll('input[type="text"], input[type="checkbox"]');
+      assert(`Expected one checkbox/radio, but ${inputs.length} found`, inputs.length === 1);
+      [input] = inputs;
+    }
 
+    not
+      ? expect(input.checked).to.be.false
+      : expect(input.checked).to.be.true;
+  },
+
+  "Then (?:the )?(?:radio button|checkbox) \"(.+?)\" should (NOT |not )?be selected in $opinionatedElement"(text, not, [collection/* , label, selector */]) {
+    assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
+    const [element] = collection;
+    const input = findInputForLabelWithText(text, element);
+
+    assert('Expected input to exist', input)
+    assert('Expected input to be a checkbox or radio button', input.type === 'checkbox' || input.type === 'radio');
+
+    not
+      ? expect(input.checked).to.be.false
+      : expect(input.checked).to.be.true;
+  },
 };
 
 export default steps;